requirements.txt
python-telegram-bot==20.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.0
psutil==5.9.5